# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HonMoon API",
    "description": "HonMoon API 문서",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/user-activities/place/{placeId}": {
      "get": {
        "tags": [
          "User Activity"
        ],
        "summary": "장소별 활동 내역 조회",
        "description": "특정 장소에서 발생한 모든 활동 내역을 조회합니다.",
        "operationId": "getActivityHistoryByPlace",
        "parameters": [
          {
            "name": "placeId",
            "in": "path",
            "description": "장소 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "currentUser",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyListUserActivityResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/missions/{id}": {
      "get": {
        "tags": [
          "Mission Detail"
        ],
        "summary": "미션 상세 정보 조회",
        "description": "특정 미션의 상세 정보를 조회합니다.",
        "operationId": "getMissionDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "미션 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "currentUser",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MissionDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/mission-places": {
      "get": {
        "tags": [
          "Mission Place"
        ],
        "summary": "미션 장소 목록 조회",
        "description": "모든 미션 장소 목록을 조회합니다.",
        "operationId": "getMissionPlaces",
        "parameters": [
          {
            "name": "currentUser",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyListMissionPlaceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/mission-places/{id}": {
      "get": {
        "tags": [
          "Mission Place"
        ],
        "summary": "미션 장소 상세 조회",
        "description": "특정 미션 장소의 상세 정보를 조회합니다.",
        "operationId": "getMissionPlace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "미션 장소 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "currentUser",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyMissionPlaceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/mission-places/{id}/missions": {
      "get": {
        "tags": [
          "Mission Place"
        ],
        "summary": "미션 장소별 미션 목록 조회",
        "description": "특정 미션 장소에 속한 모든 미션 목록을 조회합니다.",
        "operationId": "getMissionsByPlace",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "미션 장소 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          },
          {
            "name": "currentUser",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyListMissionDetailResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/mission-places/search": {
      "get": {
        "tags": [
          "Mission Place"
        ],
        "summary": "미션 장소 검색",
        "description": "장소명으로 미션 장소를 검색합니다.",
        "operationId": "searchMissionPlaces",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "description": "검색할 장소명",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "한강공원"
          },
          {
            "name": "currentUser",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyListMissionPlaceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/mission-places/nearby": {
      "get": {
        "tags": [
          "Mission Place"
        ],
        "summary": "근처 미션 장소 조회",
        "description": "사용자 위치 기준으로 근처의 미션 장소들을 조회합니다.",
        "operationId": "getNearbyMissionPlaces",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "description": "사용자 위도",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "example": 37.5665
          },
          {
            "name": "lng",
            "in": "query",
            "description": "사용자 경도",
            "required": true,
            "schema": {
              "type": "number",
              "format": "double"
            },
            "example": 126.978
          },
          {
            "name": "radius",
            "in": "query",
            "description": "검색 반경 (미터)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1000
          },
          {
            "name": "currentUser",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UserPrincipal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseBodyListMissionPlaceResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserPrincipal": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "uniqueItems": true
          }
        },
        "required": [
          "provider",
          "roles",
          "subject"
        ]
      },
...
      "AnswerCheckResult": {
        "type": "object",
        "properties": {
          "isCorrect": {
            "type": "boolean"
          },
          "confidence": {
            "type": "number",
            "format": "double"
          },
          "reasoning": {
            "type": "string"
          },
          "extractedText": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          }
        },
        "required": [
          "confidence",
          "isCorrect",
          "reasoning"
        ]
      },
      "UserActivityResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "placeId": {
            "type": "integer",
            "format": "int64"
          },
          "missionId": {
            "type": "integer",
            "format": "int64"
          },
          "description": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "isCompleted": {
            "type": "boolean"
          },
          "pointsEarned": {
            "type": "integer",
            "format": "int32"
          },
          "textAnswer": {
            "type": "string"
          },
          "selectedChoiceIndex": {
            "type": "integer",
            "format": "int32"
          },
          "uploadedImageUrl": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          },
          "aiResult": {
            "$ref": "#/components/schemas/AnswerCheckResult"
          },
          "alreadyExists": {
            "type": "boolean"
          }
        },
        "required": [
          "alreadyExists",
          "createdAt",
          "id",
          "isCompleted",
          "modifiedAt",
          "placeId",
          "pointsEarned",
          "userId"
        ]
      },
...
      "ResponseBodyListUserActivityResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserActivityResponse"
            }
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "status"
        ]
      },
...
      "MissionChoicesVo": {
        "type": "object",
        "properties": {
          "choices": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "choices"
        ]
      },
      "MissionDetailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "points": {
            "type": "integer",
            "format": "int32"
          },
          "missionType": {
            "type": "string",
            "enum": [
              "QUIZ_MULTIPLE_CHOICE",
              "QUIZ_TEXT_INPUT",
              "QUIZ_IMAGE_UPLOAD",
              "PLACE_VISIT",
              "PHOTO_UPLOAD",
              "SURVEY"
            ]
          },
          "placeId": {
            "type": "integer",
            "format": "int64"
          },
          "question": {
            "type": "string"
          },
          "answer": {
            "type": "string"
          },
          "choices": {
            "$ref": "#/components/schemas/MissionChoicesVo"
          },
          "answerExplanation": {
            "type": "string"
          },
          "correctImageUrl": {
            "type": "string"
          },
          "imageUploadInstruction": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "createdAt",
          "id",
          "missionType",
          "modifiedAt",
          "points",
          "title"
        ]
      },
      "MissionPlaceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modifiedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "createdAt",
          "id",
          "modifiedAt",
          "name"
        ]
      },
      "ResponseBodyListMissionPlaceResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MissionPlaceResponse"
            }
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "status"
        ]
      },
      "ResponseBodyMissionPlaceResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MissionPlaceResponse"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "status"
        ]
      },
      "ResponseBodyListMissionDetailResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MissionDetailResponse"
            }
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "status"
        ]
      },
...
    }
  }
}